+++ package.json
@@ -34,6 +34,7 @@
     "testee": "^0.8.1"
   },
   "dependencies": {
-    "lit-html": "^1.0.0"
+    "lit-html": "^1.0.0",
+    "lit-element": "^2.1.0"
   }
 }
