+++ src/haunted.js
@@ -1,11 +1,12 @@
-export { component } from './component.js';
-export { html, render } from './core.js';
+// export { component } from './component.js';
+export { component } from './lit-component.js';
+// export { html, render } from './core.js';
 export { useCallback } from "./use-callback.js";
 export { useEffect } from './use-effect.js';
 export { useState } from './use-state.js';
 export { useReducer } from './use-reducer.js';
 export { useMemo } from './use-memo.js';
-export { withHooks, virtual } from './virtual.js';
+// export { withHooks, virtual } from './virtual.js';
 export { useContext } from './use-context.js';
 export { createContext } from './create-context.js';
 export { hook, Hook } from './hook.js';
