+++ src/lit.js
@@ -1,1 +1,1 @@
-export * from 'https://unpkg.com/lit-html@^1.0.0/lit-html.js';
+// export * from 'https://unpkg.com/lit-html@^1.0.0/lit-html.js';
