+++ src/virtual.js
@@ -1,70 +1,70 @@
-import { Container } from './core.js';
-import { directive } from './lit.js';
-
-const partToContainer = new WeakMap();
-const containerToPart = new WeakMap();
-
-class DirectiveContainer extends Container {
-  constructor(renderer, part) {
-    super(renderer, part);
-    this.virtual = true;
-  }
-
-  commit(result) {
-    this.host.setValue(result);
-    this.host.commit();
-  }
-
-  teardown() {
-    super.teardown();
-    let part = containerToPart.get(this);
-    partToContainer.delete(part);
-  }
-}
-
-
-
-function withHooks(renderer) {
-  function factory(...args) {
-    return part => {
-      let cont = partToContainer.get(part);
-      if(!cont) {
-        cont = new DirectiveContainer(renderer, part);
-        partToContainer.set(part, cont);
-        containerToPart.set(cont, part);
-        teardownOnRemove(cont, part);
-      }
-      cont.args = args;
-      cont.update();
-    };
-  }
-
-  return directive(factory);
-}
-
-const includes = Array.prototype.includes;
-
-function teardownOnRemove(cont, part, node = part.startNode) {
-  let frag = node.parentNode;
-  let mo = new MutationObserver(mutations => {
-    for(let mutation of mutations) {
-      if(includes.call(mutation.removedNodes, node)) {
-        mo.disconnect();
-
-        if(node.parentNode instanceof ShadowRoot) {
-          teardownOnRemove(cont, part);
-        } else {
-          cont.teardown();
-        }
-        break;
-      } else if(includes.call(mutation.addedNodes, node.nextSibling)) {
-        mo.disconnect();
-        teardownOnRemove(cont, part, node.nextSibling);
-        break;
-      }
-    }
-  });
-  mo.observe(frag, { childList: true });
-}
-
-export { withHooks, withHooks as virtual }
+// import { Container } from './core.js';
+// import { directive } from './lit.js';
+//
+// const partToContainer = new WeakMap();
+// const containerToPart = new WeakMap();
+//
+// class DirectiveContainer extends Container {
+//   constructor(renderer, part) {
+//     super(renderer, part);
+//     this.virtual = true;
+//   }
+//
+//   commit(result) {
+//     this.host.setValue(result);
+//     this.host.commit();
+//   }
+//
+//   teardown() {
+//     super.teardown();
+//     let part = containerToPart.get(this);
+//     partToContainer.delete(part);
+//   }
+// }
+//
+//
+//
+// function withHooks(renderer) {
+//   function factory(...args) {
+//     return part => {
+//       let cont = partToContainer.get(part);
+//       if(!cont) {
+//         cont = new DirectiveContainer(renderer, part);
+//         partToContainer.set(part, cont);
+//         containerToPart.set(cont, part);
+//         teardownOnRemove(cont, part);
+//       }
+//       cont.args = args;
+//       cont.update();
+//     };
+//   }
+//
+//   return directive(factory);
+// }
+//
+// const includes = Array.prototype.includes;
+//
+// function teardownOnRemove(cont, part, node = part.startNode) {
+//   let frag = node.parentNode;
+//   let mo = new MutationObserver(mutations => {
+//     for(let mutation of mutations) {
+//       if(includes.call(mutation.removedNodes, node)) {
+//         mo.disconnect();
+//
+//         if(node.parentNode instanceof ShadowRoot) {
+//           teardownOnRemove(cont, part);
+//         } else {
+//           cont.teardown();
+//         }
+//         break;
+//       } else if(includes.call(mutation.addedNodes, node.nextSibling)) {
+//         mo.disconnect();
+//         teardownOnRemove(cont, part, node.nextSibling);
+//         break;
+//       }
+//     }
+//   });
+//   mo.observe(frag, { childList: true });
+// }
+//
+// export { withHooks, withHooks as virtual }
